RecommendationService/requirements.txt
flask==2.3.2
scikit-learn==1.2.2
pandas==2.0.3
numpy==1.23.5
sqlalchemy==2.1.0
psycopg2-binary==2.9.6
gunicorn==21.2.0
fastapi==0.95.2
uvicorn==0.23.2
pydantic==2.2.3
jupyterlab==4.0.5
matplotlib==3.7.2
seaborn==0.12.2
pytest==7.4.0
joblib==1.3.2